feat: S3 버킷의 프로필 이미지 제거 테스트 작성

+++ backend/src/main/java/com/darass/user/domain/SocialLoginUser.java
@@ -47,13 +47,14 @@ public class SocialLoginUser extends User {
         throw ExceptionWithMessageAndCode.NOT_GUEST_USER.getException();
     }
 
-    public void changeNickNameOrProfileImageIfExists(S3Service s3Service, String nickName,
-        MultipartFile profileImageFile) {
+    public void changeNickNameOrProfileImageIfExists(S3Service s3Service, String nickName, MultipartFile profileImageFile) {
         if (!Objects.isNull(nickName)) {
             changeNickName(nickName);
         }
         if (!Objects.isNull(profileImageFile)) {
-            s3Service.delete(getProfileImageUrl());
+            if (!getProfileImageUrl().isBlank()) {
+                s3Service.delete(getProfileImageUrl());
+            }
             String imageUrl = s3Service.upload(profileImageFile);
             changeProfileImageUrl(imageUrl);
         }
